@@ -4,7 +4,7 @@ description: testing...
 # homepage: https://www.example.com
 
 environment:
-  sdk: '>=2.10.0 <3.0.0'
+  sdk: '>=2.12.0'
 
 dependencies:
   dotenv: ^3.0.0
