@@ -1,5 +1,7 @@
 import 'package:nyxx/nyxx.dart';
 
+import 'TestBOT.dart';
+
 class Command {
   String modName;
   String cmdName;
@@ -9,6 +11,15 @@ class Command {
   Command(this.modName, this.cmdName, this.args, this.msg);
 
   TextChannel getChannel() {
-    return this.msg.channel.getFromCache();
+    return this.msg.channel.getFromCache() ??
+        terminate('There is no channel in the cache.');
+  }
+
+  String getArgumentAt(int index) {
+    if (index >= this.args.length) {
+      throw 'Invalid Argument Length';
+    } else {
+      return this.args[index];
+    }
   }
 }
